workflow: employee leave

+++ vita-framework/src/main/java/com/github/mengweijin/vita/oa/domain/entity/EmployeeLeaveDO.java
@@ -14,6 +14,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -55,7 +56,7 @@ public class EmployeeLeaveDO extends BaseEntity {
     /**
      * 请假天数
      */
-    private Integer leaveDays;
+    private BigDecimal leaveDays;
 
     /**
      * 请假原因
